requirements.txt
numpy
opencv-python
ultralytics
torchvision
torch
pillow
fastapi
python-multipart
uvicorn
